@@ -1,35 +1,25 @@
 source 'https://rubygems.org'
-ruby '2.4.1'
-gem 'rails', '5.1.1'
+ruby '2.5.1'
+gem 'rails', '5.2.0'
 
-gem 'angular-rails-templates', '~> 1.0.2'
-gem 'bcrypt', '~> 3.1.11'
-gem 'bower', '~> 0.0.5'
-gem 'jbuilder', '~> 2.6.4'
-gem 'mechanize', '~> 2.7.5'
-gem 'newrelic_rpm', '~> 4.2.0'
-gem 'nokogiri', '~> 1.7.2'
-gem 'pg', '~> 0.20.0'
-gem 'puma', '~> 3.8.2'
-gem 'rack-cors', '~> 0.4.0'
-gem 'sass-rails', '~> 5.0.6'
-gem 'sucker_punch', '~> 2.0.2'
-gem 'uglifier', '~> 3.2.0'
+gem 'bcrypt', '3.1.11'
+gem 'mechanize', '2.7.5'
+gem 'newrelic_rpm', '5.0.0.342'
+gem 'nokogiri', '1.8.2'
+gem 'pg', '1.0.0'
+gem 'puma', '3.11.4'
+gem 'rack-cors', '1.0.2'
+gem 'sucker_punch', '2.0.4'
 
 group :development do
-  gem 'debase', '~> 0.2.2.beta10'
-  gem 'reek', '~> 4.6.2'
-  gem 'rubocop', '~> 0.49.1'
-  gem 'ruby-debug-ide', '~> 0.6.0'
+  gem 'debase', '0.2.2'
+  gem 'reek', '4.8.1'
+  gem 'rubocop', '0.55.0'
+  gem 'ruby-debug-ide', '0.6.1'
 end
 
 group :test do
-  gem 'codeclimate-test-reporter', '~> 1.0.8'
-  gem 'minitest-reporters', '~> 1.1.14'
+  gem 'minitest-reporters', '1.2.0'
   gem 'shoulda', git: 'https://github.com/thoughtbot/shoulda.git', branch: 'master'
-  gem 'simplecov', '~> 0.13.0'
-end
-
-group :production do
-  gem 'rails_12factor', '0.0.3'
+  gem 'simplecov', '0.16.1'
 end
