@@ -1,4 +1,4 @@
-class ChangeTeamAbbrToNhlId < ActiveRecord::Migration
+class ChangeTeamAbbrToNhlId < ActiveRecord::Migration[4.2]
   def change
     remove_column :teams, :abbr
     add_column :teams, :nhlID, :integer
