@@ -1,5 +1,7 @@
 class StaticController < ApplicationController
   def base
-    render 'layouts/application'
+    send_data File.open("#{Rails.root}/client/dist/index.html"),
+              type: 'text/html; charset=utf-8',
+              disposition: 'inline'
   end
 end
