@@ -2,6 +2,8 @@ class Team < ApplicationRecord
   has_many :players, -> { order 'last_name, first_name' }
 
   def as_json(options = {})
-    super(options.merge(include: :players))
+    super(options.merge(
+      except: %i[created_at updated_at]
+    ))
   end
 end
